+++ src/card/sets/y1993/alpha.rs
@@ -4373,6 +4373,7 @@ pub(in crate::card::sets) static THE_HIVE: CardRecord = CardRecord::new(
         EffectDef::CreateToken {
             token: cards::WASP_TOKEN_1_1_COLORLESS,
             count: ValueDef::Constant(1),
+            tapped: false,
         },
     )),
 );