+++ src/card/sets/y1993/arabian_nights.rs
@@ -647,6 +647,7 @@ pub(in crate::card::sets) static RUKH_EGG: CardRecord = CardRecord::new(
                 effect: &EffectDef::CreateToken {
                     token: cards::BIRD_TOKEN_4_4_RED,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
             },
         )
@@ -851,6 +852,7 @@ pub(in crate::card::sets) static BOTTLE_OF_SULEIMAN: CardRecord = CardRecord::ne
 static BOTTLE_OF_SULEIMAN_WON: EffectDef = EffectDef::CreateToken {
     token: cards::DJINN_TOKEN_5_5_COLORLESS,
     count: ValueDef::Constant(1),
+    tapped: false,
 };
 
 static BOTTLE_OF_SULEIMAN_LOST: EffectDef = EffectDef::DealDamage {