+++ src/card/sets/y2012/dark_ascension.rs
@@ -94,6 +94,7 @@ pub(in crate::card::sets) static ELGAUD_INQUISITOR: CardRecord = CardRecord::new
             EffectDef::CreateToken {
                 token: cards::SPIRIT_TOKEN_1_1_WHITE,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ]),
@@ -126,6 +127,7 @@ pub(in crate::card::sets) static LINGERING_SOULS: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::SPIRIT_TOKEN_1_1_WHITE,
                 count: ValueDef::Constant(2),
+                tapped: false,
             },
         ),
         abilities::flashback(mana_cost!("{1}{B}")),
@@ -260,6 +262,7 @@ pub(in crate::card::sets) static REQUIEM_ANGEL: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::SPIRIT_TOKEN_1_1_WHITE,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ]),
@@ -574,6 +577,7 @@ pub(in crate::card::sets) static HAVENGUL_RUNEBINDER: CardRecord = CardRecord::n
                 EffectDef::CreateToken {
                     token: cards::ZOMBIE_TOKEN_2_2_BLACK,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
                 EffectDef::AddCounters {
                     object: EffectRecipientDef::MatchingObjects {
@@ -1009,6 +1013,7 @@ pub(in crate::card::sets) static REAP_THE_SEAGRAF: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::ZOMBIE_TOKEN_2_2_BLACK,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
         abilities::flashback(mana_cost!("{4}{U}")),
@@ -2123,6 +2128,7 @@ static HUNTMASTER_WOLF_AND_LIFE: EffectDef = EffectDef::Sequence(&[
     EffectDef::CreateToken {
         token: cards::WOLF_TOKEN_2_2_GREEN,
         count: ValueDef::Constant(1),
+        tapped: false,
     },
     EffectDef::GainLife {
         recipient: EffectRecipientDef::Controller,