+++ src/card/sets/y2012/return_to_ravnica.rs
@@ -227,6 +227,7 @@ static EYES_IN_THE_SKIES_EFFECTS: [EffectDef; 2] = [
     EffectDef::CreateToken {
         token: cards::BIRD_TOKEN_1_1_WHITE,
         count: ValueDef::Constant(1),
+        tapped: false,
     },
     abilities::populate(),
 ];
@@ -305,6 +306,7 @@ pub(in crate::card::sets) static KNIGHTLY_VALOR: CardRecord = CardRecord::new(
                 EffectDef::CreateToken {
                     token: cards::KNIGHT_TOKEN_2_2_WHITE,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
             ),
             AbilityDef::static_ability(
@@ -375,6 +377,7 @@ pub(in crate::card::sets) static PRECINCT_CAPTAIN: CardRecord = CardRecord::new(
                 EffectDef::CreateToken {
                     token: cards::SOLDIER_TOKEN_1_1_WHITE,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
             ),
         ]),
@@ -484,6 +487,7 @@ pub(in crate::card::sets) static SELLER_OF_SONGBIRDS: CardRecord = CardRecord::n
             EffectDef::CreateToken {
                 token: cards::BIRD_TOKEN_1_1_WHITE,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ),
@@ -1897,6 +1901,7 @@ pub(in crate::card::sets) static GOBLIN_RALLY: CardRecord = CardRecord::new(
         EffectDef::CreateToken {
             token: cards::GOBLIN_TOKEN_1_1_RED,
             count: ValueDef::Constant(4),
+            tapped: false,
         },
     )),
 );
@@ -2165,6 +2170,7 @@ pub(in crate::card::sets) static SURVEY_THE_WRECKAGE: CardRecord = CardRecord::n
             EffectDef::CreateToken {
                 token: cards::GOBLIN_TOKEN_1_1_RED,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ]),
     )),
@@ -2235,6 +2241,7 @@ pub(in crate::card::sets) static UTVARA_HELLKITE: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::DRAGON_TOKEN_6_6_RED,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ]),
@@ -2373,6 +2380,7 @@ pub(in crate::card::sets) static CENTAURS_HERALD: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::CENTAUR_TOKEN_3_3_GREEN,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ),
@@ -2520,6 +2528,7 @@ static HORNCALLERS_CHANT_EFFECTS: [EffectDef; 2] = [
     EffectDef::CreateToken {
         token: cards::RHINO_TOKEN_4_4_GREEN,
         count: ValueDef::Constant(1),
+        tapped: false,
     },
     abilities::populate(),
 ];
@@ -2747,6 +2756,7 @@ pub(in crate::card::sets) static ARMADA_WURM: CardRecord = CardRecord::new(
             EffectDef::CreateToken {
                 token: cards::WURM_TOKEN_5_5_GREEN,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ]),
@@ -2828,6 +2838,7 @@ pub(in crate::card::sets) static CALL_OF_THE_CONCLAVE: CardRecord = CardRecord::
         EffectDef::CreateToken {
             token: cards::CENTAUR_TOKEN_3_3_GREEN,
             count: ValueDef::Constant(1),
+            tapped: false,
         },
     )),
 );
@@ -3025,6 +3036,7 @@ static COURSERS_ACCORD_EFFECTS: [EffectDef; 2] = [
     EffectDef::CreateToken {
         token: cards::CENTAUR_TOKEN_3_3_GREEN,
         count: ValueDef::Constant(1),
+        tapped: false,
     },
     abilities::populate(),
 ];
@@ -3692,6 +3704,7 @@ pub(in crate::card::sets) static SELESNYA_CHARM: CardRecord = CardRecord::new(
                 EffectDef::CreateToken {
                     token: cards::KNIGHT_TOKEN_2_2_WHITE,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
             ),
         ],
@@ -4003,6 +4016,7 @@ static VRASKA_ABILITIES: [AbilityDef; 3] = [
         EffectDef::CreateToken {
             token: cards::ASSASSIN_TOKEN_1_1_BLACK,
             count: ValueDef::Constant(3),
+            tapped: false,
         },
     ),
 ];