+++ src/card/sets/y2013/dragons_maze.rs
@@ -219,6 +219,7 @@ pub(in crate::card::sets) static SUNSPIRE_GATEKEEPERS: CardRecord = CardRecord::
             EffectDef::CreateToken {
                 token: cards::KNIGHT_TOKEN_2_2_WHITE,
                 count: ValueDef::Constant(1),
+                tapped: false,
             },
         ),
     ),
@@ -923,6 +924,7 @@ pub(in crate::card::sets) static ADVENT_OF_THE_WURM: CardRecord = CardRecord::ne
         EffectDef::CreateToken {
             token: cards::WURM_TOKEN_5_5_GREEN,
             count: ValueDef::Constant(1),
+            tapped: false,
         },
     )),
 );
@@ -1709,14 +1711,17 @@ pub(in crate::card::sets) static TROSTANIS_SUMMONER: CardRecord = CardRecord::ne
                 EffectDef::CreateToken {
                     token: cards::KNIGHT_TOKEN_2_2_WHITE,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
                 EffectDef::CreateToken {
                     token: cards::CENTAUR_TOKEN_3_3_GREEN,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
                 EffectDef::CreateToken {
                     token: cards::RHINO_TOKEN_4_4_GREEN,
                     count: ValueDef::Constant(1),
+                    tapped: false,
                 },
             ]),
         ),
@@ -1806,6 +1811,7 @@ static VOICE_OF_RESURGENCE_DURING_YOUR_TURN: TriggerConditionDef =
 static VOICE_OF_RESURGENCE_TOKEN: EffectDef = EffectDef::CreateToken {
     token: cards::ELEMENTAL_TOKEN_GREEN_WHITE,
     count: ValueDef::Constant(1),
+    tapped: false,
 };
 
 // DGM 114 — Voice of Resurgence